requirements.txt
Flask
PyJWT
simplejson